Sync objective contributors on objective lift
Synchronize the objective contributor links when an objective is
lifted, as the links may be needed to keep the donor as a contributor
after the lift.
