+++ src/components/drawers/EditKeyResult.vue
@@ -163,17 +163,14 @@
 
 <script>
 import { mapState, mapGetters } from 'vuex';
-import { db } from '@/config/firebaseConfig';
-import KeyResult from '@/db/KeyResult';
-import ObjectiveContributors from '@/db/ObjectiveContributors';
 import { PktButton } from '@oslokommune/punkt-vue2';
+import { db } from '@/config/firebaseConfig';
 import { FormSection, BtnSave, BtnDelete } from '@/components/generic/form';
+import { isDepartment, isOrganization } from '@/util/getActiveItemType';
 import ArchivedRestore from '@/components/ArchivedRestore.vue';
+import KeyResult from '@/db/KeyResult';
 import PagedDrawerWrapper from '@/components/drawers/PagedDrawerWrapper.vue';
-import getActiveItemType, {
-  isDepartment,
-  isOrganization,
-} from '@/util/getActiveItemType';
+import syncObjectiveContributors from '@/util/objectiveContributors';
 
 export default {
   name: 'EditKeyResult',
@@ -332,6 +329,7 @@ export default {
   methods: {
     isDepartment,
     isOrganization,
+    syncObjectiveContributors,
 
     memberOfLevel(level) {
       return level.team.map(({ id }) => id).includes(this.user.id);
@@ -374,7 +372,7 @@ export default {
             this.thisKeyResult.objective = objectiveRef;
             this.$emit('create', this.thisKeyResult);
           }
-          await this.syncObjectiveContributor();
+          await syncObjectiveContributors(this.objective.id);
           this.$refs.drawer.next();
         } catch (error) {
           console.log(error);
@@ -390,7 +388,7 @@ export default {
       this.loading = true;
       try {
         await KeyResult.archive(this.thisKeyResult.id);
-        await this.syncObjectiveContributor();
+        await syncObjectiveContributors(this.objective.id);
         this.thisKeyResult.archived = true;
         this.$emit('archive', this.thisKeyResult);
       } catch (error) {
@@ -404,7 +402,7 @@ export default {
     async restore() {
       try {
         await KeyResult.restore(this.thisKeyResult.id);
-        await this.syncObjectiveContributor();
+        await syncObjectiveContributors(this.objective.id);
         this.thisKeyResult.archived = false;
         this.$emit('restore', this.thisKeyResult);
       } catch {
@@ -412,99 +410,6 @@ export default {
           this.$t('toaster.error.restore', { document: this.thisKeyResult.id })
         );
       }
-    },
-
-    async getObjectiveContributors() {
-      const objectiveRef = await db.doc(`objectives/${this.objective.id}`);
-      const objectiveContributors = await db
-        .collection('objectiveContributors')
-        .where('objective', '==', objectiveRef)
-        .where('archived', '==', false)
-        .get()
-        .then((snapshot) => snapshot.docs)
-        .then((docs) => docs.map((d) => d.data()));
-
-      const contributors = await Promise.all(
-        objectiveContributors.map(async (con) => {
-          return {
-            ref: await con.item.get(),
-            name: await con.item.get().then((snapshot) => snapshot.data().name),
-          };
-        })
-      );
-      return contributors;
-    },
-
-    async getKeyResultOwners() {
-      const objectiveRef = await db.doc(`objectives/${this.objective.id}`);
-      const keyResults = await db
-        .collection('keyResults')
-        .where('objective', '==', objectiveRef)
-        .where('archived', '==', false)
-        .get()
-        .then((snapshot) => snapshot.docs)
-        .then((docs) => docs.map((d) => d.data()));
-
-      const keyResultNames = await Promise.all(
-        keyResults.map(async (owner) => {
-          return {
-            ref: await owner.parent.get(),
-            name: await owner.parent.get().then((snapshot) => snapshot.data().name),
-          };
-        })
-      );
-      return keyResultNames;
-    },
-
-    async syncObjectiveContributor() {
-      const keyResultOwners = await this.getKeyResultOwners();
-      const contributors = await this.getObjectiveContributors();
-
-      let redundantContributors = [...contributors];
-      let keyResWithNoContributor = [...keyResultOwners];
-
-      // Filter out already present links
-      contributors.forEach((c) => {
-        keyResultOwners.forEach((k) => {
-          if (c.name === k.name) {
-            redundantContributors = redundantContributors.filter(
-              (con) => con.name !== c.name
-            );
-            keyResWithNoContributor = keyResWithNoContributor.filter(
-              (kr) => kr.name !== k.name
-            );
-          }
-        });
-      });
-
-      // We only need one contributor element per unique keyRes parent (here mapped by name)
-      const uniqueKeyResWithNoContributor = keyResWithNoContributor.filter(
-        (value, index, self) => index === self.findIndex((t) => t.name === value.name)
-      );
-
-      // Add missing contributor
-      uniqueKeyResWithNoContributor.forEach((k) => {
-        this.createObjectiveContributor(k.ref);
-      });
-
-      // Remove redundant contributors
-      redundantContributors.forEach((c) => {
-        this.removeObjectiveContributor(c.ref);
-      });
-    },
-
-    createObjectiveContributor(item) {
-      const itemType = getActiveItemType(item.data());
-      const itemRef = db.doc(`${itemType}s/${item.id}`);
-      const objectiveRef = db.doc(`objectives/${this.objective.id}`);
-      ObjectiveContributors.create(itemRef, objectiveRef);
-    },
-
-    async removeObjectiveContributor(item) {
-      const itemType = getActiveItemType(item.data());
-      const itemRef = db.doc(`${itemType}s/${item.id}`);
-      const objectiveRef = db.doc(`objectives/${this.objective.id}`);
-      return ObjectiveContributors.remove(itemRef, objectiveRef);
     },
 
     close(e) {
